fix/uint256=uint

+++ contracts/2key/acquisition-campaign-contracts/TwoKeyConversionHandler.sol
@@ -53,8 +53,7 @@ contract TwoKeyConversionHandler is Upgradeable, TwoKeyConversionStates, TwoKeyC
      */
     uint [] counters;
 
-    uint256 expiryConversionInHours; // How long converter can be pending before it will be automatically rejected and funds will be returned to convertor (hours)
-
+    uint expiryConversionInHours; // How long converter can be pending before it will be automatically rejected and funds will be returned to convertor (hours)
 
     address twoKeyEventSource;
     address contractor;
